backend/requirements.txt
fastapi
uvicorn
numpy
opencv-python
joblib
imdbpy
python-multipart
python-dotenv
